Added lots of logic for the season system
This is a large commit which should maybe have been split up into multiple ones, but I couldn't bother.
It adds the basis for the season system with some features implemented and some still missing.
The major missing features are lack of snow in the winter and lack of persistence between saves.